adistetsa/requirements.txt
asgiref==3.4.1
astroid==2.9.0
attrs==21.4.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.5
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.18.7
django-adminactions==1.15
django-cleanup==6.0.0
django-colorfield==0.6.3
django-config-models==2.3.0
django-crum==0.7.9
django-debug-toolbar==3.2.4
django-etc==1.3.1
django-extensions==3.1.5
django-filebrowser==4.0.1
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==3.0.2
django-import-export==2.7.1
django-subadmin==3.2.1
django-waffle==2.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
edx-django-utils==4.5.0
et-xmlfile==1.1.0
factory-boy==3.2.1
factory-djoy==2.2.0
Faker==10.0.0
fcm-django==1.0.12
filelock==3.4.2
filetype==1.0.13
firebase-admin==5.3.0
Flask==2.0.0
fpdf==1.7.2
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
html-json-forms==1.1.1
httplib2==0.20.4
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itertable==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.2
lazy-object-proxy==1.7.1
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.4
natural-keys==2.0.0
newrelic==7.4.0.172
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pbr==5.8.1
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.6
psutil==5.9.0
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydotplus==2.0.2
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==7.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.9
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
swagger-spec-validator==2.7.4
tablib==3.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.1.2
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
